requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pprint==0.1
pymongo==3.10.1
python-dotenv==0.10.5
pytz==2019.3
six==1.14.0
Werkzeug==0.16.1
